client: consolidate adaptive startup code into a test helper
