+++ client/main_test.go
@@ -1,15 +1,11 @@
 package client
 
 import (
-	"context"
-	"fmt"
-	"os"
 	"testing"
-	"time"
 
 	"github.com/tendermint/tendermint/abci/example/kvstore"
 	nm "github.com/tendermint/tendermint/node"
-	rpctest "github.com/tendermint/tendermint/rpc/test"
+	"github.com/tendermint/tendermint/rpc/test"
 )
 
 // useful values for test cases
@@ -29,28 +25,7 @@ func TestMain(m *testing.M) {
 
 	// run the default kvstore app inside a tendermint instance
 	app := kvstore.NewKVStoreApplication()
-	fmt.Println("Starting tendermint...")
-	node = rpctest.StartTendermint(app)
-
-	// make sure tendermint is good to go before tests... a short static pause,
-	// then wait for one block to come in.
-	fmt.Println("Wait for first block...")
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	h, err := NewClient(NewLocalConnection(node)).WaitForNextBlock(ctx)
-	fmt.Printf("Starting tests with block %d\n", h.Height)
-
-	// Run tests if tendermint started properly
-	var code int
-	if err == nil {
-		code = m.Run()
-	} else {
-		fmt.Printf("Failed to start tendermint: %s\n", err)
-		code = 1
-	}
-
-	// and shut down proper at the end
-	node.Stop()
-	node.Wait()
-	os.Exit(code)
+	TestWithTendermint(app, func(n *nm.Node) {
+		node = n
+	}, m)
 }
